+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.3 (Mar 23, 2020)
+
+* Added 3 new battle cards: **Bloodletting**, **Bodyguard**, and **Critical Strike**.
+* Added a unlock pack to the mod so it can be enabled and disabled in game. Comes after the 10th pack in each character.
+* Buffed **Preach**: 3 -> 4 **indoctrination** for unupgraded card, and 5 -> 6 for enhanaced card.
+
 ## 0.0.2 (Mar 21, 2020)
 
 * Change the card set to neutral so it works with Rook(Probably).